Fix privileges for sidecar injector's mutatingwebhookconfiguration
The sidecar injector currently lists _all_ mutatingwebhookconfigurations
instead of only getting and watching the configured one.

Additionally, using a resourceName for the create verb isn't allowed by
Kubernetes. If you want to allow a user to create a resource, the user
needs to have a general create privilege (one that doesn't specify
resourceNames).