MAISTRA-425 Add PodLocality controller
Istio needs to know the zone and region of each pod. Since that
information is not stored on the Pod, but on the Node object, it would
need read privileges to the cluster-scoped Node object.

The idea of the PodLocality controller is to copy the region and zone
labels from the Node to the Pod, allowing Istio to get the info directly
from the Pod.