[MAISTRA-625] Enable CNI unconditionally on OCP 4
We check for the presence of the NetworkAttachmentDefinition CRD and enable CNI
for all Control Planes in the cluster if it is present.

If it's not present, then we disable CNI, likely in OCP 3 environments.

This means that we are overriding whatever the user puts in the `istio_cni.enabled`
field of the SMCP definition.