Remove README.md reference to controller
This repo is for software which leverages the [operator framework][1] in
a single application focus. While in the most technical sense it _is_ a
controller (insomuch as all operators are controllers, but not all
controllers are operators) as we extend the API through the use of a
custom resource definition it's best to err on the side of brevity
(unlike this commit message).

[1]: https://github.com/operator-framework/getting-started