[MAISTRA-624] Use basic HTTP auth for internal communication
Between kiali, grafana, jaeger and prometheus.

This way we can get rid of cluster role bindings.

We create an htpasswd secret unconditionally on the beginning of the control
plane installation, configure all oauth proxies to use that htpasswd
file as another form of authentication and finally configure Kiali
CR to use HTTP basic auth to communicate with those services.