MAISTRA-726 Handle SMCP disappearing from cache during removal of finalizer
If the SMCP is deleted, the reconciler deletes resources owned by the
SMCP. This requeues the same SMCP, causing the reconcile to be invoked
again immediately after it has completely deleted the SMCP (i.e. removed
its finalizer). Because the watch event caused by the deletion may
arrive during the invocation of reconcile(), the SMCP object may
disappear from the cache mid-way through. The update will fail with a
conflict, and the get() on the cache will then fail with a NotFound
error. If that happens, reconcile() can safely return without any errors.

If there's any other error during get(), there's no sense in calling
get() again immediately, as that will most likely result in the same
error. This will just result in a flood of errors in the log. Instead,
reconcile() now returns an error, properly requeueing the SMCP with
backoff.