MAISTRA-746 Set maxUnavailable, maxSurge for grafana, prometheus, ior and gateway deployments
Istio 1.1 is still using the extensions/v1beta1 API definition of
deployments, which defaults maxUnavailable to 1. This can lead to the
operator assuming the deployments are available, when in fact they are
not, e.g. due to one of the containers in a pod not being able to start. This
commit injects a rollingUpdate strategy into deployments that don't yet
define it.