Don't delete the ServiceMeshMemberRoll if it's already deleted
After deleting the SMMR, it still exists until the MemberRoll controller
performs the clean up. In the mean time, the Member controller may be
triggered again and will delete the SMMR again. This isn't problematic,
since the deletion operation is idempotent, but we do get multiple
"Removing ServiceMeshMember from ServiceMeshMemberRoll" statements in the
log. This commit fixes this.