MAISTRA-1500 Fix race condition in post-install readiness check (#447)
After deploying a component, if the cache isn't updated in time, we never
wait for the component to become ready (its has no entry in the
readinessMap; we consider this to mean that the chart has no resources
with readiness information, so we consider the whole chart as ready).

If the component is missing in readinessMap, we must take that as a signal
that the component is NOT ready. But if we do that, we'll take all components
that have no resources with readiness information as not ready.

So, before checking readiness, we should determine if the chart has any
resources that contain readiness information. If it doesn't we should
count the component as ready.

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>