Configure mergify to add "ready to merge" label to PRs
This is to make it easy for us to see which PRs have been approved
and are only missing the "okay to merge" label.

This should also make it easy to see which PRs still need at least
one additional approval (PRs that aren't marked as WIP and have no
"ready to merge" label).

I experimented with also adding a "needs review" label, but then
the PR discussion would have a bunch of 'added/removed "needs review"'
events, because switching from WIP to nonWIP and approvals & their
revocations would cause the label to be added/removed.