MAISTRA-1897: Disable ingressClassName support
Istio now supports determining whether to process Ingress resources by
reading the new (as of Kubernetes v1.18) `ingressClassName` field in
their spec, as opposed to only checking the deprecated annotation.
The `IngressClass` resources are cluster-scoped however.  An option to
disable this was added in maistra/istio#188 -- this updates the charts
to use it.