MAISTRA-1938 Ensure all ports are forwarded to the proxy by default
When using the init container, all ports are forwarded, but the
template for CNI is such that iptables rules are configured to
only forward the ports that are specified in the pod definition,
while all others go straight to the application container.

This causes the AuthorizationPolicy to not have any effect on
these other ports.