MAISTRA-2460 Only log "ignoring control plane namespace" when it's in SMMR.spec.members
namespacesToReconcile is always sourced from SMMR.spec.members, whereas namespacesToRemove is determined by listing namespaces with the member-of label. The control plane namespace contains this label and is therefore always in the namespacesToRemove list. We thus shouldn't log the warning "ignoring control plane namespace in members list of ServiceMeshMemberRoll" in that case.