MAISTRA-2486 Fix flaky test TestMAISTRA_2040 (#763)
Since verifyActions.React() in verifiers.go expects the controller to perform actions in the order specified in the test, the test shouldn't test the update of both the validating and the mutating webhook, as the order in which these two resources are updated by the controller isn't deterministic. The test must therefore be split into two.