MAISTRA-2475 Fix initialization and update of Kiali.spec.deployment.accessible_namespaces (#777)
This ensures that the .spec.deployment.accessible_namespaces field is initialized when the Kiali resource is first deployed (by the smcp controller), but is then updated only by the smmr controller. The smcp controller should never overwrite the value, as this would cause the issue described in MAISTRA-1956. Previously, we simply removed the field from the Helm chart, which resulted in the smcp controller never updating the field (it was thus only ever updated by the smmr controller), but this resulted in a different issue, where Kiali starts logging errors, because it thinks it has access to all namespaces.