Prevent deadlock during operator update
When a rolling update of the operator is performed, the new pod never becomes ready because it doesn't start the webhook server until it becomes the leader. It can't become the leader until the old pod terminates. But in a rolling update, the old pod isn't terminated until the new pod becomes ready.