MAISTRA-2236: Make NetworkPolicy optional
This adds the boolean flag `spec.security.manageNetworkPolicy` to
the SMCP, with the default value of `true`.

If it is true, current behavior is preserved, i.e., the operator creates
and manages the `NetworkPolicies` that allow communication between the
control plane and namespaces that are members of the mesh.

If it is false, the operator doesn't create such resources. The user is
entirely responsible for managing the `NetworkPolicies` and troubleshoot
any issue that might occur as a result of opting out of this feature.