Make scorecard test work in kind (#1663)
Previously, it relied on a k8s cluster to be setup in advance, which worked for us
because this job used to run in an OCP cluster, therefore the cluster was always
there, pointed by `$KUBECONFIG`.

In order to migrate to Istio Prow we need to run this job in a kind cluster, hence this PR.

Side benefit is that now we can run the test locally directly with `make test.scorecard`.