Removes error logging for expected failures, adds requeues (#1670)
Previously, we were stack tracing quite regularly in circumstances that
are to be expected, especially in a two-controller setup like we have.

This also adds a requeue (with info log) for the rare case where we see
the RESTMapper of the gc admission plugin being out of date. This mostly
happened in CI - I wasn't able to reproduce locally, even after 50
attempts.