Fix integration tests
Recursive make is evil.

Currently the `IMAGE` variable is read at the first invocation of
make, and is set to the default value of `quay.io/maistra-dev/sail-operator:3.0-latest`.
This value is not being changed between nested invocations ok make, thus
tests are failing because this image doesn't exist. We only figured
about this because a recent commit changed the default value from
`maistra-dev` to `sail-dev`. We are effectively testing against an
already pushed image (likely from yesterday). We are supposed to use the
image we just built and pushed to the internal registry.

This patch fixes this by not using recursive make - thus some target
recipes are kind of duplicated. I tried to put them in a macro but,
again, recursive make is evil and bad things happen.

We can improve it in a follow up, but for now this should be sufficient
to unblock developers and CI.