OSSM-6700 Fix rest_client_request_latency_seconds metrics
Due to a bug in kubernetes/client-go, the generated metrics were too course. All the URLs that the kube rest client was calling were collapsed into "https://apiserver/%7Bprefix%7D". This made these metrics completely useless, since they put everything in the same bucket.

The bug in client-go was fixed much later (not in any 0.18.x release), so we had to fork the client-go repository and patch the code by copying it from the master branch. The patch is in github.com/maistra/kubernetes-client-go, branch v0.18.3-patch.

With this fix, the metrics are now correctly split across different URLs. Here are some examples:

- https://apiserver/api/v1/configmaps
- https://apiserver/api/v1/endpoints
- https://apiserver/api/v1/namespaces
- ...