OSSM-6697 Fix duplicate caCertConfigMapName env var
As part of #1035 I had to move the addition of the environment variable into the
maistra/istio repository and forgot to update the patch script in the operator repo.
With out the move, the field would not have been available
during integration tests. In the past we usually made all chart changes
in the patch script, but then we had to reimplement parts of that to
simulate OSSM behavior in the maistra/istio integration tests, which I
wanted to avoid.